Moved fotogrid code to its own file. Created subfolder for static css.
