+++ templates/template.html
@@ -3,8 +3,8 @@
   <head>
     <!-- CSS -->
     {{ fontawesome_html() }}
-    <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
-    <link rel="fotogrid stylesheet" href="{{ url_for('static', filename='fotogrid.css') }}">
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}">
+    <link rel="fotogrid stylesheet" href="{{ url_for('static', filename='css/fotogrid.css') }}">
     <link rel="shortcut icon" href="{{ url_for('static', filename='favicon.ico') }}">
   </head>
   <body class=post>
